requirements.txt
asgiref
Django
djangorestframework
djangorestframework-simplejwt
PyJWT
python-dotenv
sqlparse
